BE on typescript. Separate secrets for new jwt and rooms
