--- common/interfaces/index.ts
@@ -1,18 +1,0 @@
-
-export interface User {
-    id?: string;
-    name?: string;
-    avatarUrl?: string;
-    device?: string;
-
-}
-
-export interface AuthInfo {
-    token: string;
-    user: User;
-}
-
-export interface RoomInfo {
-    roomId: string;
-    roomUsers: User[];
-}
