+++ package.json
@@ -7,19 +7,23 @@
     "license": "MIT",
     "private": true,
     "scripts": {
-        "start": "tsc && node dist/server.js",
-        "start-js": "node server/server.js",
+        "start": "node dist/server.js",
         "start-dev": "concurrently --kill-others-on-fail \"yarn server\" \"yarn client\"",
-        "server": "NODE_ENV=development nodemon server/server.js",
+        "build": "tsc",
+        "build-ui": "cd client && yarn && yarn build",
+        "server": "NODE_ENV=development nodemon server/server.ts",
         "client": "cd client && yarn start",
-        "heroku-postbuild": "cd client && yarn && yarn build"
+        "heroku-postbuild": "yarn build-ui"
     },
     "devDependencies": {
         "@types/compression": "^1.7.2",
         "@types/express": "4.17.1",
+        "@types/jsonwebtoken": "^8.5.8",
         "@types/passport": "^1.0.7",
+        "@types/passport-jwt": "^3.0.6",
         "concurrently": "^5.1.0",
-        "nodemon": "^2.0.2",
+        "nodemon": "^2.0.15",
+        "ts-node": "^10.7.0",
         "tslint": "^6.1.3",
         "typescript": "^4.6.2"
     },
