+++ tsconfig.json
@@ -7,7 +7,9 @@
     "moduleResolution": "node",
     "sourceMap": true,
     "outDir": "dist",
-    "strict": true
+    "strict": true,
+    "removeComments": true,
+    "preserveConstEnums": true
   },
   "lib": ["es2015"],
   "exclude": [
